+++ grpckit/__init__.py
@@ -1,5 +1,5 @@
 from .service import Service  # noqa: F401
 from .app import GrpcKitApp  # noqa: F401
-from ._global import current_app, g, request  # noqa: F401
+from .globals import current_app, g, request  # noqa: F401
 
 __version__ = "0.1.0"